+++ api/models/comment/createAndPresentComment.js
@@ -1,7 +1,8 @@
 import { sanitize } from 'hylo-utils/text'
 import { difference, uniq } from 'lodash'
 import { simpleUserColumns } from '../../presenters/UserPresenter'
-import { normalizeComment } from '../../../lib/util/normalize'
+import { normalizeComment, normalizedSinglePostResponse } from '../../../lib/util/normalize'
+import { postRoom, pushToSockets, userRoom } from '../../services/Websockets'
 
 export function validateCommentCreateData (userId, data) {
   return Post.isVisibleToUser(data.postId, userId)
@@ -96,11 +97,33 @@ const presentComment = comment =>
     return Object.assign(buckets, c)
   })
 
-const notifySockets = (comment, post) => {
+function notifySockets (comment, post) {
   if (post.get('type') === Post.Type.THREAD) {
     const followerIds = post.relations.followers.pluck('id')
-    return post.pushMessageToSockets(comment, followerIds)
-  } else {
-    return post.pushCommentToSockets(comment)
+    return pushMessageToSockets(post, comment, followerIds)
   }
+
+  return pushCommentToSockets(post, comment)
+}
+
+// n.b.: `message` has already been formatted for presentation
+function pushMessageToSockets (thread, message, userIds) {
+  const excludingSender = userIds.filter(id => id !== message.user_id.toString())
+
+  if (thread.get('num_comments') === 0) {
+    return Promise.map(excludingSender, userId => {
+      const opts = {withComments: 'all'}
+      return this.load(PostPresenter.relations(userId, opts))
+      .then(post => PostPresenter.present(post, userId, opts))
+      .then(normalizedSinglePostResponse)
+      .then(post => pushToSockets(userRoom(userId), 'newThread', post))
+    })
+  }
+
+  return Promise.map(excludingSender, id =>
+    pushToSockets(userRoom(id), 'messageAdded', {postId: thread.id, message}))
+}
+
+function pushCommentToSockets (post, comment) {
+  return pushToSockets(postRoom(post.id), 'commentAdded', {comment, postId: post.id})
 }
